feat(P1-009): consolidate test utilities in browser-common/src/test
- Create consolidated setupTests.ts with enhanced mocks and JSDOM support
- Move testUtils.ts to proper location in browser-common
- Update jest configurations to use shared test setup
- Fix TextEncoder/TextDecoder polyfill issues with CommonJS imports
- Ensure both browser-common and js packages use unified test environment
- All existing tests continue to pass with improved mock capabilities

Test plan:
✅ Test setup works in new location
✅ Mock utilities function correctly
✅ No test runner issues
✅ Build succeeds for all packages
✅ All unit tests pass in both packages
