+++ packages/browser-common/jest.config.js
@@ -3,6 +3,6 @@ module.exports = {
   transform: {
     '\\.[jt]sx?$': ['babel-jest', { configFile: './../../babel.config.js' }],
   },
-  setupFiles: ['./src/setupTests.ts'],
+  setupFiles: ['./src/test/setupTests.ts'],
   testMatch: ['<rootDir>/src/**/*.test.ts'],
 }
