+++ packages/js/jest.config.js
@@ -4,6 +4,6 @@ module.exports = {
     '\\.[jt]sx?$': ['babel-jest', { configFile: './../../babel.config.js' }],
   },
   resolver: '<rootDir>/test/resolver.js',
-  setupFiles: ['./src/setupTests.ts'],
+  setupFiles: ['../browser-common/src/test/setupTests.ts'],
   testMatch: ['<rootDir>/src/**/*.test.ts'],
 }
